re-organized routes to follow rosetta-sdk-go pattern
