+++ crates/mentat-server/src/server/mod.rs
@@ -23,28 +23,25 @@ use crate::{api::*, conf::*};
 /// instance of [`Server`] using [`ServerBuilder`], or with the
 /// [`super::mentat`] macro if a default instance using your custom types is
 /// preferred.
-///
-/// ```ignore
-/// struct MentatBitcoin;
-/// impl ServerType for MentatBitcoin {
-///     type CallApi = call_api::BitcoinCallApi;
-///     type ConstructionApi = construction_api::BitcoinConstructionApi;
-///     type DataApi = data_api::BitcoinDataApi;
-///     type IndexerApi = indexer_api::BitcoinIndexerApi;
-///     type CustomConfig = node::NodeConfig;
-/// }
-/// ```
 pub trait ServerType: Sized + 'static {
+    /// The blockchain's `AccountApi` Rosetta implementation.
+    type AccountApi: AccountApiRouter;
+    /// The blockchain's `BlockApi` Rosetta implementation.
+    type BlockApi: BlockApiRouter;
     /// The blockchain's `CallApi` Rosetta implementation.
-    type CallApi: CallerCallApi;
+    type CallApi: CallApiRouter;
     /// The blockchain's `ConstructionApi` Rosetta implementation.
-    type ConstructionApi: CallerConstructionApi;
-    /// The blockchain's `CallerDataApi` Rosetta implementation.
-    type DataApi: CallerDataApi;
-    /// The blockchain's `IndexerApi` Rosetta implementation.
-    type IndexerApi: CallerIndexerApi;
+    type ConstructionApi: ConstructionApiRouter;
+    /// The blockchain's `Events` Rosetta implementation.
+    type EventsApi: EventsApiRouter;
+    /// The blockchain's `MempoolApi` Rosetta implementation.
+    type MempoolsApi: MempoolApiRouter;
+    /// The blockchain's `NetworkApi` Rosetta implementation.
+    type NetworkApi: NetworkApiRouter;
     /// Any optional endpoints for the Mentat implementation.
-    type OptionalApi: OptionalApi;
+    type OptionalApi: OptionalApiRouter;
+    /// The blockchain's `SearchApi` Rosetta implementation.
+    type SearchApi: SearchApiRouter;
     /// The nodes's `NodeConf` implementation.
     type CustomConfig: DeserializeOwned + NodeConf;
 
@@ -90,16 +87,24 @@ pub trait ServerType: Sized + 'static {
 
 /// The Struct for building a Server.
 pub struct ServerBuilder<Types: ServerType> {
-    /// The optional API endpoints.
-    optional_api: Option<Types::OptionalApi>,
-    /// The call API endpoints.
+    /// The Account API endpoints.
+    account_api: Option<Types::AccountApi>,
+    /// The Block API endpoints.
+    block_api: Option<Types::BlockApi>,
+    /// The Call API endpoints.
     call_api: Option<Types::CallApi>,
-    /// The construction API endpoints.
+    /// The Construction API endpoints.
     construction_api: Option<Types::ConstructionApi>,
-    /// The data API endpoints.
-    data_api: Option<Types::DataApi>,
-    /// The indexer API endpoints.
-    indexer_api: Option<Types::IndexerApi>,
+    /// The Events API endpoints.
+    events_api: Option<Types::EventsApi>,
+    /// The Mempool API endpoints.
+    mempool_api: Option<Types::MempoolsApi>,
+    /// The network API endpoints.
+    network_api: Option<Types::NetworkApi>,
+    /// The search API endpoints.
+    search_api: Option<Types::SearchApi>,
+    /// The Optional API endpoints.
+    optional_api: Option<Types::OptionalApi>,
     /// The optional configuration details.
     configuration: Option<Configuration<Types::CustomConfig>>,
 }
@@ -107,12 +112,16 @@ pub struct ServerBuilder<Types: ServerType> {
 impl<Types: ServerType> Default for ServerBuilder<Types> {
     fn default() -> Self {
         Self {
-            optional_api: None,
+            account_api: None,
+            block_api: None,
             call_api: None,
-            configuration: None,
             construction_api: None,
-            data_api: None,
-            indexer_api: None,
+            events_api: None,
+            mempool_api: None,
+            network_api: None,
+            search_api: None,
+            optional_api: None,
+            configuration: None,
         }
     }
 }
@@ -121,19 +130,65 @@ impl<Types: ServerType> ServerBuilder<Types> {
     /// Builds the Server.
     pub fn build(self) -> Server<Types> {
         Server {
-            optional_api: self
-                .optional_api
-                .expect("You did not set the additional api."),
+            account_api: self.account_api.expect("You did not set the call api."),
+            block_api: self.block_api.expect("You did not set the call api."),
             call_api: self.call_api.expect("You did not set the call api."),
-            configuration: self
-                .configuration
-                .expect("You did not set the custom configuration."),
             construction_api: self
                 .construction_api
                 .expect("You did not set the construction api."),
-            data_api: self.data_api.expect("You did not set the data api."),
-            indexer_api: self.indexer_api.expect("You did not set the indexer api."),
+            events_api: self.events_api.expect("You did not set the call api."),
+            mempool_api: self.mempool_api.expect("You did not set the call api."),
+            network_api: self.network_api.expect("You did not set the call api."),
+            optional_api: self
+                .optional_api
+                .expect("You did not set the additional api."),
+            search_api: self.search_api.expect("You did not set the call api."),
+            configuration: self
+                .configuration
+                .expect("You did not set the custom configuration."),
         }
+    }
+
+    /// Sets the Account API on the builder.
+    pub fn account_api(mut self, a: Types::AccountApi) -> Self {
+        self.account_api = Some(a);
+        self
+    }
+
+    /// Sets the Block API on the builder.
+    pub fn block_api(mut self, a: Types::BlockApi) -> Self {
+        self.block_api = Some(a);
+        self
+    }
+
+    /// Sets the Call API on the builder.
+    pub fn call_api(mut self, a: Types::CallApi) -> Self {
+        self.call_api = Some(a);
+        self
+    }
+
+    /// Sets the construction API on the builder.
+    pub fn construction_api(mut self, a: Types::ConstructionApi) -> Self {
+        self.construction_api = Some(a);
+        self
+    }
+
+    /// Sets the Events API on the builder.
+    pub fn events_api(mut self, a: Types::EventsApi) -> Self {
+        self.events_api = Some(a);
+        self
+    }
+
+    /// Sets the Mempool API on the builder.
+    pub fn mempool_api(mut self, a: Types::MempoolsApi) -> Self {
+        self.mempool_api = Some(a);
+        self
+    }
+
+    /// Sets the Network API on the builder.
+    pub fn network_api(mut self, a: Types::NetworkApi) -> Self {
+        self.network_api = Some(a);
+        self
     }
 
     /// Sets the optional API on the builder.
@@ -142,9 +197,9 @@ impl<Types: ServerType> ServerBuilder<Types> {
         self
     }
 
-    /// Sets the call API on the builder.
-    pub fn call_api(mut self, a: Types::CallApi) -> Self {
-        self.call_api = Some(a);
+    /// Sets the Search API on the builder.
+    pub fn search_api(mut self, a: Types::SearchApi) -> Self {
+        self.search_api = Some(a);
         self
     }
 
@@ -165,51 +220,45 @@ impl<Types: ServerType> ServerBuilder<Types> {
         self.configuration = Some(Configuration::load(path));
         self
     }
-
-    /// Sets the construction API on the builder.
-    pub fn construction_api(mut self, a: Types::ConstructionApi) -> Self {
-        self.construction_api = Some(a);
-        self
-    }
-
-    /// Sets the data API on the builder.
-    pub fn data_api(mut self, a: Types::DataApi) -> Self {
-        self.data_api = Some(a);
-        self
-    }
-
-    /// Sets the indexer API on the builder.
-    pub fn indexer_api(mut self, a: Types::IndexerApi) -> Self {
-        self.indexer_api = Some(a);
-        self
-    }
 }
 
 /// The server struct for running the Rosetta server.
 pub struct Server<Types: ServerType> {
-    /// The optional API endpoints.
-    pub optional_api: Types::OptionalApi,
-    /// The call API endpoints.
+    /// The Account API endpoints.
+    pub account_api: Types::AccountApi,
+    /// The Block API endpoints.
+    pub block_api: Types::BlockApi,
+    /// The Call API endpoints.
     pub call_api: Types::CallApi,
-    /// The configuration API endpoints.
-    pub configuration: Configuration<Types::CustomConfig>,
-    /// The construction API endpoints.
+    /// The Construction API endpoints.
     pub construction_api: Types::ConstructionApi,
-    /// The data API endpoints.
-    pub data_api: Types::DataApi,
-    /// The indexer API endpoints.
-    pub indexer_api: Types::IndexerApi,
+    /// The Events API endpoints.
+    pub events_api: Types::EventsApi,
+    /// The Mempool API endpoints.
+    pub mempool_api: Types::MempoolsApi,
+    /// The network API endpoints.
+    pub network_api: Types::NetworkApi,
+    /// The search API endpoints.
+    pub search_api: Types::SearchApi,
+    /// The Optional API endpoints.
+    pub optional_api: Types::OptionalApi,
+    /// The optional configuration details.
+    pub configuration: Configuration<Types::CustomConfig>,
 }
 
 impl<Types: ServerType> Default for Server<Types> {
     fn default() -> Self {
         Self {
-            optional_api: Default::default(),
+            account_api: Default::default(),
+            block_api: Default::default(),
             call_api: Default::default(),
-            configuration: Types::CustomConfig::load_config(),
             construction_api: Default::default(),
-            data_api: Default::default(),
-            indexer_api: Default::default(),
+            events_api: Default::default(),
+            mempool_api: Default::default(),
+            network_api: Default::default(),
+            search_api: Default::default(),
+            optional_api: Default::default(),
+            configuration: Types::CustomConfig::load_config(),
         }
     }
 }
